implemented skip login and fixed login
got screen to properly appear, despite not being able to actually login, although doesnt seem like our fault. this could potentially be an API key that we need to actually login. we may try to get one later but i also introduced a skip button for the demo. all of the functions for login are in place so it might not be necessary for a prototype.